Refactor Rubric into multiple files
